Append vaccine parameters from nimue model

This calls parameters from nimue and appends the following parameters to the list:

gamma_vaccine_delay: mean duration of period from vaccination to vaccine protection.

gamma_V: mean duration of vaccine-derived immunity (days)

rel_infectiousness_vaccinated: Relative infectiousness per age category of vaccinated individuals relative to unvaccinated individuals. Default = rep(1, 17), which is no impact of vaccination on onwards transmissions

rel_infectiousness: Relative infectiousness per age category relative to maximum infectiousness category. Default = rep(1, 17)

vaccine_efficacy_infection: Efficacy of vaccine against infection. This parameter must either be length 1 numeric (a single efficacy for all age groups) or length 17 numeric vector (an efficacy for each age group). An efficacy of 1 will reduce FOI by 100 percent, an efficacy of 0.2 will reduce FOI by 20 percent etc.

prob_hosp: Efficacy of vaccine against severe (requiring hospitilisation) disease (by age). This parameter must either be length 1 numeric (a single efficacy for all age groups) or length 17 numeric vector (an efficacy for each age group). An efficacy of 1 will reduce the probability of hospitalisation by 100 percent, an efficacy of 0.2 will reduce the probability of hospitalisation by 20 percent etc.

vaccine_coverage_mat: Vaccine coverage targets by age (columns) and priority (row)

N_prioritisation_steps: number of vaccine prioritization levels

current_prioritisation_step: current prioritization step we are on

vaccine_set: vaccines available each day of simulation

append_vaccine_nimue(parameters, ...)

Arguments

parameters
list from get_parameters

...
Other parameters for nimue:::parameters