Get parameters for safir (nimue vaccine model)

This calls parameters from nimue.

get_parameters_nimue(
  population,
  contact_mat,
  time_period,
  max_age = 100,
  lambda_external = NULL,
  dt = 1,
  ...
)

Arguments

population

integer vector of population by age group

contact_mat

a contact matrix from get_mixing_matrix

time_period

maximum time of simulation

max_age

max age for humans

lambda_external

a vector of length equal to time_period giving an additional additive term contributing to the force of infection. Models infectious contacts the population has with external (unmodeled) populations.

dt

time step size

...

Other parameters for nimue::parameters