Vaccination distribution process (multiple doses, no types)

This function distributes vaccine doses each day for the vaccine model with multiple doses but no types. It distributes vaccines according to the following strategy.

* start at \code{phase = 1}, meaning we begin by distributing the first dose
* use \code{\link{get_vaccination_priority_stage}} to check what stage of the
prioritization matrix \code{vaccine_coverage_mat} from \code{\link[nimue]{strategy_matrix}}
we should be using to decide who to vaccinate.
* we can advance one row of the prioritization matrix when coverage for this dose (phase) >=
prioritization matrix and coverage for next dose (phase) >= prioritization matrix groups prioritized
for the next dose by the matrix \code{next_dose_priority}.

vaccination_process(parameters, variables, events, dt)

Arguments

parameters
a named list

variables
a named list

events
a named list

dt
size of time step